CNN_Example_MNIST
=================================== EXPLANATION OF MODEL =================================

The shape of the data is 28x28 pixel greyscale images (Channel = 1). The output size with a kernel_size = 5, stride = 2, and padding of 2 will be 32 channels of 14x14 images (down sampled by factor of 2 due to the max_pooling). 

For the first layer of the model class, we will use nn.Sequential() container to list the objects. First, the convolutional layer (nn.Conv2d) will have an input channel size of 1 (greyscale) and an output channel size of 32. Next, the activation function (nn.ReLU()) will be a rectified linear unit. Finally, the last object of the first layer is a Max_Pool function (nn.MaxPool2d()) with a stride of 2 (to downs-sample the size of the input by factor of 2), and kernel_size of 2.

The second layer of the model class is defined the same way except that the input_channel size for the conv2d object is 32 (the size of the output_channels from the first layer) and the size of the output channels for the second layer would be 64. Since we are using the same max_pooling kernel_size and strides, the second layer will also be down sized by a factor of 2 to 64 channels of 7x7 pixel images. 

We specify a dropout layer to avoid over-fitting the model.

For the fully connected layers, we use nn.Linear(). The first parameter (input_size) will be the number of channels/nodes from the previous layers multiplied by the length and width of the output image from previous layers. In this case, we have 64 out_channels from the 64 filters and output image size of 7x7 due to the downsampling of factor of 2 in the second layer's max_pool. So the flattened tensor for the input of the linear layer would be 64 x 7 x 7 (3184). The output size would be 1000 nodes, which will be the input size for the second linear layer.


================================== TRAINING THE MODEL ====================================

We have two loops: we iterate over the number of epochs and also iterate over the length of train_loader set using enumerate. This allows us to pass the images of the MNIST data set from train_loader through the forward pass of the model to calculate the prediction of the model.

Next, pass the model prediction and the TRUE IMAGES through the CrossEntropyLoss function, defined as criterion. We created a total_loss list, which we will append to as we gather loss after each epoch. 

Then, we perform backward propagation and optimization of the model parameters using ADAM. First we zero the gradients using .zero_grad() and then calculate gradients on the loss variable using .backward(). Finally, using the gradients calculated in the back-prop, we call optimizer.step() to perform the ADAM optimizer and update the model parameters. 

Finally, we want to keep track of the model accuracy by using torch.max() which will return the index of the maximum value in a tensor. The arguments of this function are the tensor that will be observed and the axis over which to determine the maximum. 

Because the output size of the last fully connected linear layer in the model is size 10, the output of the model will have a size (batch_size, 10). 

To determine the prediction, we need to find the maximum value over the 10 output nodes, each of which will correspond to one of the handwritten digits (i.e output #2 will correspond to digit '2').
The output node with the highest value will be the final prediction (i.e if output #2 has a value of 1.0 (maximum) then the predicted digit will be 2).

This is why we set the second argument of the torch.max() function to 1 - 1 will be the highest value possible for any output. 

After returning a list of prediction digits, we will compare the predicted digits with the true images (predicted == label). == in Python is a COMPARISON OPERATOR which simply means “If the values of two operands are equal, then the condition becomes true”. We sum them to determine how many correct predictions there are. *Output of the .sum() is still a tensor!*

To find the accuracy, the we divide the number of correct predictions by the batch_size, which is equal to labels.size(). 

We print the progress after every 100 epochs.


==================================== TESTING THE MODEL ======================================

First, set the model to eval() mode to stop the dropout and use torch.no_grad() to stop gradient descent/optimization. Next, iterate over test loader and compute the accuracy by dividing the sum of the correct predictions over the total number of true labels. 

*There are 100 labels in the test_loader so we will have 100 iterations. Each iteration will have 10,000 images from the test_dataset*

Next, save the model using torch.save() function. 

Finally, plot the accuracy and loss.